requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
openai>=1.0.0
anthropic>=0.18.0
google-genai>=0.1.0
tenacity>=8.2.0
python-dotenv>=1.0.0
httpx>=0.24.0
supabase>=2.3.0
e2b>=0.14.0

